Refactor OutputFacade to meet new services
