+++ detekt-migration/src/main/kotlin/io/gitlab/arturbosch/detekt/migration/Migration.kt
@@ -22,7 +22,7 @@ class Migration {
 				val settings = ProcessingSettings(inputPath, config, createPathFilters(), parallel, true)
 				val detektor = Detektor(settings, KtTreeCompiler.instance(settings), listOf(MigrationRuleSetProvider()))
 				val detektion = detektor.run()
-				OutputFacade(this, detektion).consoleFacade()
+				OutputFacade(this, detektion, settings).run()
 			}
 
 		}
